added test cases for service and repo layer

+++ src/main/java/com/demo/Service/InventoryService.java
@@ -14,53 +14,82 @@ public class InventoryService {
     InventoryRepo repo=new InventoryRepo();
     PrintUtility printUtility=new PrintUtility();
 
-    public void addInventory(String itemName, String sellingPrice, String buyPrice) {
+    public boolean addInventory(String itemName, String sellingPrice, String buyPrice) {
 
-        repo.addInventory(new Inventory(itemName,convertToDouble(sellingPrice),convertToDouble(buyPrice),0));
-
+        return repo.addInventory(new Inventory(itemName,convertToDouble(sellingPrice),convertToDouble(buyPrice),0));
     }
 
-    public void updateBuyInventory(String itemName, int quantity) {
+    public boolean updateBuyInventory(String itemName, String boughtQuantity) {
 
         Inventory inv= repo.getInventoryByName(itemName);
-        quantity+=inv.getQuantity();
-        inv.setQuantity(quantity);
-        repo.updateInventory(inv);
+        int quantity=convertToInt(boughtQuantity);
+        if(inv==null)
+            return false;
+        else {
+            quantity += inv.getQuantity();
+            inv.setQuantity(quantity);
+            repo.updateInventory(inv);
+            return true;
+        }
     }
 
-    public void updateSellInventory(String itemName, int quantity){
+    public boolean updateSellInventory(String itemName, String soldQuantity){
 
         Inventory inv =repo.getInventoryByName(itemName);
+        int quantity=convertToInt(soldQuantity);
         //update profit
-        Report.profit+=(inv.getSellingPrice()-inv.getBuyPrice())*quantity;
-        quantity=inv.getQuantity()-quantity;
-        inv.setQuantity(quantity);
-        repo.updateInventory(inv);
+        if(inv == null)
+            return false;
+        else {
+            quantity = inv.getQuantity() - quantity;
+            if(quantity<0){
+                return false;
+            }
+            else {
+                Report.profit += (inv.getSellingPrice() - inv.getBuyPrice()) * quantity;
+                inv.setQuantity(quantity);
+                repo.updateInventory(inv);
+                return true;
+            }
+        }
     }
 
-    double convertToDouble(String num){
+    private double convertToDouble(String num){
         double d=0;
-        try {
-            d=Double.parseDouble(num);
-        }catch(NumberFormatException e){
-            System.out.println("Invalid decimal");
+        if(num!=null) {
+            d = Double.parseDouble(num);
         }
         return d;
     }
 
-    public void report() {
+    public boolean report() {
 
         //print inventory
         Set<Inventory> inventories=repo.getAllInventories();
         printUtility.printReport(inventories);
         //reset profit
-        Report.profit=0;
+        Report.profit = 0;
+        return true;
     }
 
-    public void delete(String itemName) {
+    public boolean delete(String itemName) {
         Inventory inv = repo.getInventoryByName(itemName);
-        //update profit
-        Report.profit-=inv.getBuyPrice()*inv.getQuantity();
-        repo.delete(inv);
+        if(inv==null)
+            return false;
+        else {
+            //update profit
+            Report.profit -= inv.getBuyPrice() * inv.getQuantity();
+            repo.delete(inv);
+            return true;
+        }
+    }
+    private int convertToInt(String num){
+        int i=0;
+        try {
+            i= Integer.parseInt(num);
+        }catch(NumberFormatException e){
+            System.out.println("Invalid number format");
+        }
+        return i;
     }
 }
